# Supabase-Gateway/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SahatGhar Telemedicine Admin Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /support/tickets:
    get:
      operationId: listSupportTickets
      tags: [support]
      summary: List all support tickets
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: user_type
          in: query
          schema:
            type: string
        - name: assigned_to
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Tickets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketListResponse"
  /support/tickets/{id}:
    get:
      operationId: getSupportTicket
      tags: [support]
      summary: Get a support ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ticket detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"

    patch:
      operationId: updateSupportTicket
      tags: [support]
      summary: Update a support ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketUpdate"
      responses:
        "200":
          description: Updated ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"
...
components:
  schemas:
...
    SupportTicket:
      type: object
      required: [id, ticket_id, category, status, priority, raised_by, date, user_type]
      properties:
        id:
          type: string
        ticket_id:
          type: string
        category:
          type: string
        status:
          type: string
        priority:
          type: string
        raised_by:
          type: string
        user_type:
          type: string
        date:
          type: string
        assigned_to:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        resolution_notes:
          type: ["string", "null"]
        linked_appointment_id:
          type: ["string", "null"]
        refund_amount:
          type: ["number", "null"]
    TicketListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/SupportTicket"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    TicketUpdate:
      type: object
      properties:
        status:
          type: string
        assigned_to:
          type: string
        resolution_notes:
          type: string
        priority:
          type: string
...